junk/pyproject.toml
[project]
name = "junk"
version = "0.0.1"
authors = [
  { name="Madison Brooks", email="maddie@0xmaddie.page" },
]
description = "🗑️"
readme = "readme.md"
requires-python = ">=3.7"
dependencies = [
  "pycairo",
  "numpy",
  "scipy",
  "opensimplex",
  "loguru",
  "sentencepiece",
  "transformers",
  "accelerate",
  "bitsandbytes",
  "ftfy",
  # "playwright",
  "diffusers @ git+https://github.com/huggingface/diffusers",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
